Add files via upload
This the Sleeping Professor problem. It is consists of a Professor process and a process for each student. 
The office has a waiting room with n chairs for waiting studets. If a student is in the waiting room and there are no free chairs, the students will leave.
If a student enters and the waiting room is empty. The student must wait untile another student to wake up the professor. Only two students can wake up the professor. The professor will check for waiting students when the two students leave. 
The professor will invide student to his office and answer questions, otherwise he goes to sleep in his office.